+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-29T15:15:11.668Z"
+    created: Optional[str] = "2026-07-30T03:03:04.599Z"
 
-    updated: Optional[str] = "2026-07-29T15:15:11.668Z"
+    updated: Optional[str] = "2026-07-30T03:03:04.599Z"
 
     updated_by_id: OptionalNullable[str] = UNSET